Improve names for duplicated elements
When adding a elements with a name that already exists in the out file,
add an underline before the appended index.

This fixes a confusion when adding few inputs with names such as "R1" and
"R11". Before this commit it was hard to seperate between duplicated "R1"
and pre-existing "R11". Those names are used to identfy the components in
the output of some ULP scripts.